Continue implementing structs

+++ dust-lang/src/jit_vm/register.rs
@@ -13,7 +13,7 @@ pub union Register {
     pub integer: i64,
     pub prototype_index: usize,
     pub object_pointer: *mut Object,
-    pub register_indices: (u16, u16),
+    pub register_indices: (u32, u32),
 }
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
